refactor(storage): move HistoryInfo to historical.rs
Move HistoryInfo enum and history_info_from_shard function from
either_writer.rs to their semantic home in historical.rs where
the history lookup logic belongs.
